drivers: soc/mfd: fixup compilation/warning

+++ drivers/soc/samsung/exynos-cpu_hotplug.c
@@ -142,8 +142,8 @@ static struct cpumask create_cpumask(void)
 	int cpu;
 	struct cpumask mask;
 
-	online_cpu_min = min(pm_qos_request(PM_QOS_CPU_ONLINE_MIN), nr_cpu_ids);
-	online_cpu_max = min(pm_qos_request(PM_QOS_CPU_ONLINE_MAX), nr_cpu_ids);
+	online_cpu_min = min((int)pm_qos_request(PM_QOS_CPU_ONLINE_MIN), (int)nr_cpu_ids);
+	online_cpu_max = min((int)pm_qos_request(PM_QOS_CPU_ONLINE_MAX), (int)nr_cpu_ids);
 
 	cpumask_clear(&mask);
